Sort fragments during operation normalization (De-composed from #1027). (#1112)
* Cutover to apollo-graphql

* Add apollo-graphql as a dependency (and project reference)
* Remove apollo-engine-reporting as a dependency
* Update sortAST to normalize order of fragments w.r.t operations
* Tests are failing expectedly at this point

* Update tests

This update is exactly as expected. All fragments will now move to the beginning of a normalization result.

* Use snapshot tests

This is a valid use case for snapshots. Even better would be inline snapshots - something we can get to later.

* Centralize operation hashing function

These two operations will likely be used in tandem, and we want this to be consistent across consumers.

* Incorporate rename suggestions

* Add explicit comment about sort order

* Transform the operation in a less aggressive manner - don't remove aliases and be less strict about removing literals.

* Update incorrect reference to renamed module, update related snapshots.

* Revert changes to defaultEngineReportingSignature. Apply changes to new function, defaultOperationRegistrySignature.

This new function is the effective interim fix, and the current existing function is now left alone.

* Add CHANGELOG.md entry for #1112.
