+++ packages/apollo-graphql/src/index.ts
@@ -1,1 +1,5 @@
-export { defaultEngineReportingSignature } from "./signature";
+export {
+  defaultEngineReportingSignature,
+  defaultOperationRegistrySignature,
+  operationHash
+} from "./operationId";
