+++ packages/apollo-graphql/src/transforms.ts
@@ -101,6 +101,14 @@ function sorted<T>(
 // order, it can make sure the queries are treated as identical.
 export function sortAST(ast: DocumentNode): DocumentNode {
   return visit(ast, {
+    Document(node: DocumentNode) {
+      return {
+        ...node,
+        // Use sortBy because 'definitions' is not optional.
+        // The sort on "kind" places fragments before operations within the document
+        definitions: sortBy(node.definitions, "kind", "name.value")
+      };
+    },
     OperationDefinition(
       node: OperationDefinitionNode
     ): OperationDefinitionNode {
